requirements.txt
numpy
pandas
mysql-connector-python
pyodbc
requests
chardet